+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Feature
 
 - Implemented check if note is in the correct folder already bevor move
+- Implemented drag&drop reordering for rules and filters
 
 ### Bug Fixes